# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PersonaFit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/surveys": {
      "post": {
        "tags": [
          "surveys"
        ],
        "summary": "Create Survey",
        "operationId": "create_survey_api_surveys_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "surveys"
        ],
        "summary": "List Surveys Endpoint",
        "operationId": "list_surveys_endpoint_api_surveys_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "running",
                    "completed",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/surveys/{survey_id}/run": {
      "post": {
        "tags": [
          "survey_run"
        ],
        "summary": "Trigger Run",
        "description": "설문 시뮬레이션 시작 (백그라운드).\n\nArgs:\n    force: True이면 status='running'이라도 강제로 다시 시작.\n           pm2 재시작 등으로 백그라운드 작업이 끊겨 세션이 stale running으로 남았을 때 복구용.\n           completed 세션은 보존(캐시 hit으로 즉시 통과), running/failed 세션은 pending으로 리셋.\n\nReturns:\n    {status: \"started\", survey_id, total, reset, completed_preserved}",
        "operationId": "trigger_run_api_surveys__survey_id__run_post",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Run Api Surveys  Survey Id  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionConfig": {
        "properties": {
          "llm_provider": {
            "type": "string",
            "enum": [
              "anthropic",
              "sllm"
            ],
            "title": "Llm Provider",
            "default": "anthropic"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "claude-haiku-4-5"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "include_reasoning": {
            "type": "boolean",
            "title": "Include Reasoning",
            "default": true
          }
        },
        "type": "object",
        "title": "ExecutionConfig",
        "description": "LLM 실행 파라미터."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "uuid4"
          },
          "order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Order",
            "description": "1-based 노출 순서"
          },
          "type": {
            "type": "string",
            "enum": [
              "single_choice",
              "multi_choice",
              "scale",
              "open_ended",
              "nps"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Options"
          },
          "scale_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Min"
          },
          "scale_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Max"
          },
          "scale_label_low": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Label Low"
          },
          "scale_label_high": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Label High"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "order",
          "type",
          "text"
        ],
        "title": "Question",
        "description": "설문 1문항.\n\n유형별 필드 사용:\n  - single_choice / multi_choice: options 필수 (최소 2개)\n  - scale: scale_min/scale_max 필수 (예: 1-5, 1-7)\n  - open_ended: 추가 필드 없음\n  - nps: scale_min=0, scale_max=10 고정"
      },
...
      "Survey": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "title": "Objective",
            "description": "조사 목적 (LLM 응답 품질 향상에 활용)",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "default": "draft"
          },
          "target_filter": {
            "$ref": "#/components/schemas/TargetFilter"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionConfig"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "persona_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Persona Uuids",
            "description": "Step 2 확정 시점의 페르소나 스냅샷"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "target_filter",
          "execution",
          "created_at",
          "updated_at"
        ],
        "title": "Survey",
        "description": "설문 본체. 마법사 Step 1~4의 결과물."
      },
      "SurveyCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "title": "Objective",
            "default": ""
          },
          "target_filter": {
            "$ref": "#/components/schemas/TargetFilter"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionConfig"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "persona_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Persona Uuids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "target_filter",
          "execution"
        ],
        "title": "SurveyCreateRequest"
      },
      "SurveyListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "SurveyListResponse"
      },
...
      "TargetFilter": {
        "properties": {
          "age_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Min"
          },
          "age_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Max"
          },
          "sex": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sex"
          },
          "provinces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provinces"
          },
          "family_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Family Types"
          },
          "education_levels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Education Levels"
          },
          "occupations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Occupations"
          },
          "query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "sampling": {
            "type": "string",
            "enum": [
              "all",
              "random_n",
              "proportional"
            ],
            "title": "Sampling",
            "default": "random_n"
          },
          "sample_size": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Sample Size",
            "default": 100
          }
        },
        "type": "object",
        "title": "TargetFilter",
        "description": "페르소나 선별 조건 — /api/dataset/personas/filter 입력과 호환."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}